backend/requirements.txt
# Core
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1

# MCP
mcp>=1.0.0

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP client
httpx==0.27.2

# Google Calendar
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.153.0

# Email
sendgrid==6.11.0

# Notifications
slack-sdk==3.33.4
twilio==9.3.5

# LLM
openai==1.54.4
anthropic==0.39.0

# Python misc
python-multipart==0.0.12
